pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "waverecox-professional"
version = "2.0.0"
description = "Enhanced RF Spectrum Analysis Tool"
authors = [{name = "WaveReconX Team"}]
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "numpy>=1.21.0",
    "pandas>=1.3.0",
    "scipy>=1.7.0",
    "requests>=2.25.0",
    "urllib3>=1.26.0",
    "json5>=0.9.5",
    "yaml>=5.4.1",
    "structlog>=21.1.0",
    "asyncio>=3.4.3",
    "pathlib>=1.0.1",
    "watchdog>=2.1.0",
    "cryptography>=3.4.8",
    "pytz>=2021.1",
    "openpyxl>=3.0.7",
    "regex>=2021.4.4",
    "typing-extensions>=3.10.0",
    "matplotlib>=3.4.0",
    "plotly>=5.0.0",
    "seaborn>=0.11.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.2.5",
    "pytest-cov>=2.12.0",
    "black>=21.5b2",
    "flake8>=3.9.0",
    "mypy>=0.812",
    "sphinx>=4.0.0",
    "sphinx-rtd-theme>=0.5.2",
]

sdr = [
    "pyrtlsdr>=0.2.93",
    "scapy>=2.4.5",
    "dpkt>=1.9.8",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.isort]
profile = "black"
multi_line_output = 3
